requirements.txt
Django==1.9.5
PyJWT==1.4.0
SQLAlchemy==1.0.12
argparse==1.4.0
behave==1.2.5
diff-match-patch==20121119
dj-config-url==0.1.1
dj-static==0.0.6
django-custom-user==0.5
django-import-export==0.4.5
django-pipeline==1.6.8
django-pipeline-browserify==0.2.0
django-postgrespool==0.3.0
django-webpack-loader==0.3.0
djangorestframework==3.3.2
djangorestframework-jwt==1.8.0
enum34==1.1.2
extras==1.0.0
factory-boy==2.6.1
fake-factory==0.5.7
fixtures==3.0.0
funcsigs==1.0.2
fuzzywuzzy==0.11.0
gunicorn==19.4.5
ipaddress==1.0.16
linecache2==1.0.0
mock==2.0.0
parse==1.6.6
parse-type==0.3.4
pbr==1.10.0
psycopg2==2.6.1
python-dateutil==2.5.2
python-mimeparse==1.5.2
python-subunit==1.2.0
selenium==2.53.1
six==1.10.0
splinter==0.7.3
static3==0.6.1
sure==1.4.0
tablib==0.11.2
testtools==2.2.0
traceback2==1.4.0
unittest2==1.1.0
wheel==0.24.0
whitenoise==3.0
wsgiref==0.1.2
